requirements.txt
# Core scraping libraries
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
playwright==1.40.0
scrapy==2.11.0

# Data processing and storage
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
pymongo==4.6.0
redis==5.0.1

# Async and concurrency
aiohttp==3.9.1
asyncio-throttle==1.0.2
concurrent-futures==3.1.1

# Utilities
fake-useragent==1.4.0
python-dotenv==1.0.0
schedule==1.2.0
urllib3==2.1.0
urllib-robotparser==1.0.0

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0

# CAPTCHA solving
2captcha-python==1.1.3
anticaptchaofficial==1.0.46

# Proxy management
requests-ip-rotator==1.0.14
proxy-randomizer==0.1.0

# Data validation
cerberus==1.3.4
jsonschema==4.20.0

# Export formats
openpyxl==3.1.2
xlsxwriter==3.1.9

# Image processing (for CAPTCHA)
Pillow==10.1.0
opencv-python==4.8.1.78

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0